% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_crowding_distance.R
\name{calculate_crowding_distance}
\alias{calculate_crowding_distance}
\title{计算前沿的拥挤距离}
\usage{
calculate_crowding_distance(front)
}
\arguments{
\item{front}{列表，前沿粒子子列表}
}
\value{
列表，已添加"crowding_distance"的前沿粒子
}
\description{
计算前沿的拥挤距离
}
