% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_archive.R
\name{manage_archive}
\alias{manage_archive}
\title{管理外部存档（限制大小，使用拥挤距离）}
\usage{
manage_archive(archive, max_size)
}
\arguments{
\item{archive}{列表，外部存档粒子}

\item{max_size}{正整数，最大存档大小}
}
\value{
列表，截取后的存档
}
\description{
管理外部存档（限制大小，使用拥挤距离）
}
