% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gantt_chart.R
\name{plot_gantt_chart}
\alias{plot_gantt_chart}
\title{Plot Gantt chart for scheduling results}
\usage{
plot_gantt_chart(
  schedule,
  title = "Gantt Chart",
  filename = "gantt_chart.png",
  width = 14,
  height = 8,
  dpi = 300
)
}
\arguments{
\item{schedule}{Result$schedule returned by decode_schedule()}

\item{title}{Plot title}

\item{filename}{File name only. Default "gantt_chart.png"; NULL means not saved.}

\item{width}{Image width (inches)}

\item{height}{Image height (inches)}

\item{dpi}{Resolution}
}
\description{
Plot Gantt chart for scheduling results
}
