% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/example_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Example dataset for MOPSO task scheduling}
\format{
A named list-like collection of scheduling-related objects

A collection of example objects used for MOPSO experiments.
}
\source{
Simulated example data

Simulated example data
}
\usage{
data(test_data)

data(test_data)
}
\description{
This dataset contains all example objects required to run the MOPSO algorithm,
including task definitions and resource configurations.

This dataset contains all example data required to run the MOPSO algorithm,
including task definitions, resource information, and cost parameters.
}
\details{
The dataset contains the following objects:
\describe{
\item{tasks}{List. Task definitions including activities, priorities, and precedence.}
\item{self_resource}{List. In-house resource definitions.}
\item{outsourced_resource}{List. Outsourced resource definitions.}
\item{resource_windows}{List. Resource availability time windows.}
\item{resource_joint_cost}{List. Joint cost between resources.}
\item{resource_mapping}{List. Mapping between in-house and outsourced resources.}
}

After calling \code{data(test_data)}, the following objects become available:
\describe{
\item{tasks}{Task definitions}
\item{self_resource}{In-house resource information}
\item{outsourced_resource}{Outsourced resource information}
\item{resource_windows}{Resource availability windows}
\item{resource_joint_cost}{Joint resource cost matrix}
\item{resource_mapping}{Task–resource mapping}
}
}
